Admin ux improve (#835)
* refactor: update component logic, styling, and hierarchy
- extracted the side panel logic to another component in an attempt to
  make it reusable (not only used for searching)

* refactor: update components for admin and map controls
- change location for displaying suggested changes by putting it in the
  sidepanel
- sidepanel is more versatile since instead of checking for if-else
  statements of other stores, it checks for type
- changed toggle of the controls by changing the button text instead
- changed map rotation in dimension toggling by removing the bearing
  change logic
- created a new store type for the sidepanel component
- although incomplete, the map flies to the recommend location (although
  markers are still missing)

* fix(side-panel): reopen the panel from the browse helpers and drop debug alert

The new SidePanelStore only renders while `active` is set, and `active` is set
only by `openPanel()`. `openCampusBrowse` / `openBrowseClasses` had lost their
`expand()` call without gaining an `openPanel()`, so every caller that routes
through them set a query and rendered nothing — no error, no failing test.

- Both helpers now call `openPanel()` themselves, so all call sites are fixed at
  once instead of hand-pairing `updateQuery` + `openPanel` at each one.
- `openBrowseClasses`'s `sidePanelStore` param is optional, so Sidebar (which
  opens the panel itself) and CampusBrowseChips both keep working untouched.
- `ClassesList` was aliased to CampusBrowseList.svelte; point it at the real file.
- Remove the leftover `alert("event marker clicked")` in Map.svelte; the
  `openPanel()` call right above it is the intended behaviour.

Regression tests use the real SidePanelStore (a double implementing only
`expand()` kept passing while the panel stayed invisible) and assert the call
sites still pass the store.

* fix(side-panel): resolve panel content from the query, not openPanel alone

The side panel rendered only while `sidePanelStore.active` was set, and
`active` was set only by `openPanel()`. Deep links, browser back/forward and
every caller that just sets a query never reach `openPanel()`, so `/room/…`,
`/building/…` and `/event/…` loaded to an empty map.

`resolvePanelContent()` keeps `openPanel()` metadata as the override and falls
back to the query category, so all paths render from one gate instead of
pairing `openPanel()` at each call site. The review queue, which has no query
category, still comes from the metadata.

Also restores what the rewrite dropped: the jeepney stop and route panels, the
`aria-hidden` on collapsed panel content, and the mobile collapsed render gate.

* fix(side-panel): rebuild the panel on staging's bottom sheet and drop openPanel pairing

The redesign replaced the hand-rolled mobile drag with BottomSheet and moved
the 2D/3D control into MapControlsStack, so the panel this PR extracted was
carrying a superseded copy of both.

- SidePanel now uses staging's BottomSheet on mobile and staging's drawer
  chrome on desktop, keeping resolvePanelContent as the single gate so deep
  links still render.
- MainControlsStore folded into SidePanelStore, so mobileSheetSnap survives.
- The 2D toggle's bearing reset is dropped in MapControlsStack, the live
  control, instead of the now-orphaned MapDimensionToggle.
- Sidebar lets the browse helpers open the panel instead of pairing
  openPanel at the call site; the call-site guard covers openBrowseClasses.

* docs(browse): correct the stale gate comment on the browse helpers

---------
