+++ src/components/svelte/EditorShelf.svelte
@@ -14,6 +14,7 @@
     mapEditStore,
     modalStore,
     proposalsStore,
+    sidePanelStore,
     toastStore,
   } from "@lib/store.svelte";
   import { beginEventPlacement } from "@lib/event-placement";
@@ -75,7 +76,8 @@
 
   function handleReviewQueue() {
     onclose?.();
-    modalStore.openModal("review");
+    sidePanelStore.openPanel({ type: "admin-suggestions" });
+    sidePanelStore.expand();
   }
 
   function addEventLabel() {
