+++ src/components/svelte/Entry.svelte
@@ -401,6 +401,7 @@
           locationStore.clearDestination();
         }
       }
+      sidePanelStore.closePanel();
     }
   }
 </script>
