+++ src/components/svelte/Map.svelte
@@ -166,6 +166,12 @@
     submitCreateProposal,
     submitPinPositionProposal,
   } from "@lib/proposals/client";
+  import BuildingResult from "./controls/BuildingResult.svelte";
+  import PlaceResult from "./controls/PlaceResult.svelte";
+  import OrgResult from "./controls/OrgResult.svelte";
+  import DormResult from "./controls/DormResult.svelte";
+  import EventResult from "./controls/EventResult.svelte";
+
   const data = getAppData();
   const appActions = getAppActions();
   const { buildings, dorms, events, organizations, places, loaded } =
@@ -282,7 +288,10 @@
       value: place.name,
     });
     queryStore.inputValue = place.name;
-    sidePanelStore.expand();
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: PlaceResult,
+    });
   }
 
   // Event titles are not unique, so resolve the selected event by its slug when
@@ -2564,7 +2573,7 @@
     });
   });
 
-  function handleMarkerClick(buildingName: string) {
+  function handleBuildingMarkerClick(buildingName: string) {
     if (eventPlacementStore.active) return;
     if (isMapEditEnabled() && selectedEditKey !== null) return;
     if (buildingName === queryStore.inputValue) return;
@@ -2574,6 +2583,10 @@
       value: buildingName,
     });
     queryStore.inputValue = buildingName;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: BuildingResult,
+    });
   }
 
   function handleDormMarkerClick(dormName: string) {
@@ -2586,6 +2599,10 @@
       value: dormName,
     });
     queryStore.inputValue = dormName;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: DormResult,
+    });
   }
 
   function handleOrgMarkerClick(name: string) {
@@ -2604,7 +2621,10 @@
       value: name,
     });
     queryStore.inputValue = name;
-    sidePanelStore.expand();
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: OrgResult,
+    });
   }
 
   function handleEventMarkerClick(event: EventData) {
@@ -2618,6 +2638,10 @@
       eventSlug: event.slug,
     });
     queryStore.inputValue = event.title;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: EventResult,
+    });
   }
 
   function toggleEventMarkerGroup(groupKey: string) {
@@ -3335,7 +3359,7 @@
                 <Marker
                   lngLat={[position.lon, position.lat]}
                   draggable={canDragPin(editKey)}
-                  onclick={() => handleMarkerClick(building.buildingName)}
+                  onclick={() => handleBuildingMarkerClick(building.buildingName)}
                   ondragstart={() => beginMarkerDrag(editKey)}
                   ondragend={(e) =>
                     handleBuildingDragEnd(
