+++ src/components/svelte/controls/CampusBrowseList.svelte
@@ -15,7 +15,18 @@
     isLandmarkPlaceCategory,
     placeDirectoryLabel,
   } from "@constants/place-categories";
-  import { jeepneyStore, queryStore, transitStore } from "@lib/store.svelte";
+  import {
+    jeepneyStore,
+    queryStore,
+    sidePanelStore,
+    transitStore,
+  } from "@lib/store.svelte";
+  import CollegeResult from "./CollegeResult.svelte";
+  import BuildingResult from "./BuildingResult.svelte";
+  import DivisionResult from "./DivisionResult.svelte";
+  import OrgResult from "./OrgResult.svelte";
+  import DormResult from "./DormResult.svelte";
+  import PlaceResult from "./PlaceResult.svelte";
   import { campusTransit } from "../../../campus.config";
 
   // Derived from campusTransit.label so a fork edits one place:
@@ -299,6 +310,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: BuildingResult,
+    });
   }
 
   function openCollege(name: string) {
@@ -308,6 +323,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: CollegeResult,
+    });
   }
 
   function openDivision(name: string) {
@@ -317,6 +336,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: DivisionResult,
+    });
   }
 
   function openOrg(name: string) {
@@ -326,6 +349,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: OrgResult,
+    });
   }
 
   function openDorm(name: string) {
@@ -335,6 +362,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: DormResult,
+    });
   }
 
   function openPlace(name: string) {
@@ -344,6 +375,10 @@
       value: name,
     });
     queryStore.inputValue = name;
+    sidePanelStore.openPanel({
+      type: "search-result",
+      component: PlaceResult,
+    });
   }
 
   const filteredJeepneyRoutes = $derived.by(() => {
@@ -364,6 +399,7 @@
 
   function closeList() {
     queryStore.clearQuery();
+    sidePanelStore.closePanel();
   }
 
   function onFilterInput(event: Event) {
