+++ src/components/svelte/editor/review-panel.css
@@ -1,9 +1,18 @@
+/* Scoped to the panel: EntityReviewActions imports this file from a module,
+   which makes it a global stylesheet, so a bare `*` reset would land on every
+   page. */
+.entity-review-panel * {
+  box-sizing: border-box;
+}
+
 .entity-review-panel {
   display: flex;
   flex-direction: column;
   gap: 0.375rem;
   border-top: 1px solid hsl(0, 0%, 92%);
   padding-top: 0.4375rem;
+  height: 100%;
+  box-sizing: border-box;
 }
 
 .entity-review-heading {
@@ -35,8 +44,8 @@
   display: flex;
   flex-direction: column;
   gap: 0.4375rem;
-  max-height: 16rem;
   overflow-y: auto;
+  flex: 1 0 auto;
 }
 
 .entity-review-item {
@@ -44,6 +53,14 @@
   border-radius: 0.625rem;
   padding: 0.4375rem;
   background: hsl(0, 0%, 99%);
+  display: block;
+  width: 100%;
+  &:hover {
+    border: 1px solid hsl(0, 70%, 80%);
+  }
+  &:focus-within {
+    border: 1px solid hsl(0, 70%, 32%);
+  }
 }
 
 .entity-review-meta {
