+++ src/lib/stores/ui-stores.svelte.ts
@@ -8,6 +8,7 @@ import type {
   QueryStoreState,
   RecentSearch,
   SidebarOpenType,
+  SidePanelMetaData,
 } from "./store-types";
 
 export class ModalStore {
@@ -178,27 +179,6 @@ export class ToastStore {
 /** Mobile entity bottom sheet snap; desktop stays `"closed"`. */
 export type MobileSheetSnap = "closed" | "peek" | "expanded";
 
-export class MainControlsStore {
-  collapsed: boolean = $state(false);
-  mobileSheetSnap: MobileSheetSnap = $state("closed");
-
-  toggle = () => {
-    this.collapsed = !this.collapsed;
-  };
-
-  expand = () => {
-    this.collapsed = false;
-  };
-
-  collapse = () => {
-    this.collapsed = true;
-  };
-
-  setMobileSheetSnap = (snap: MobileSheetSnap) => {
-    this.mobileSheetSnap = snap;
-  };
-}
-
 export class FloatingControlPanelStore {
   openPanel: FloatingControlPanel | null = $state(null);
 
@@ -244,5 +224,38 @@ export class SidebarStore {
 
   closeRail = () => {
     this.railOpen = false;
+  };
+}
+
+export class SidePanelStore {
+  state = $state<SidePanelMetaData | null>(null);
+  active = $state<boolean>(false);
+  collapsed: boolean = $state(false);
+  mobileSheetSnap: MobileSheetSnap = $state("closed");
+
+  openPanel(state: SidePanelMetaData) {
+    this.state = state;
+    this.active = true;
+  }
+
+  closePanel = () => {
+    this.active = false;
+    this.state = null;
+  };
+
+  toggleVisibility = () => {
+    this.collapsed = !this.collapsed;
+  };
+
+  expand = () => {
+    this.collapsed = false;
+  };
+
+  collapse = () => {
+    this.collapsed = true;
+  };
+
+  setMobileSheetSnap = (snap: MobileSheetSnap) => {
+    this.mobileSheetSnap = snap;
   };
 }
